move cdoXml block positioning logic to serialization helpers
