+++ apps/src/blockly/addons/cdoXml.js
@@ -59,106 +59,8 @@ export default function initializeBlocklyXml(blocklyWrapper) {
       });
     });
 
-    // Below this point, we will be repositioning blocks on a rendered workspace.
-    if (!blockSpace.rendered) {
-      return blocks;
-    }
-
-    // Once we are always converting XML to JSON, we can remove all of this repositioning logic.
-    // Note that RTL languages position blocks from the left within a
-    // blockSpace. For instructions and embedded hints, there is no viewWidth,
-    // so we determine the starting point based on the width of the block.
-    const metrics = blockSpace.getMetrics();
-    const viewWidth = metrics ? metrics.viewWidth : 0;
-    const blockWidth = blocks[0]
-      ? blocks[0].blockly_block.getHeightWidth().width
-      : 0;
-    // Add padding if we are in a workspace.
-    const padding = viewWidth ? 16 : 0;
-    const verticalSpaceBetweenBlocks = 10;
-
-    // The cursor is used to position blocks that don't have explicit x/y coordinates
-    let cursor = {
-      x: padding,
-      y: padding,
-    };
-    if (blockSpace.RTL && viewWidth) {
-      // Position the cursor from the right of the workspace.
-      cursor.x = viewWidth - padding;
-    } else if (blockSpace.RTL && !viewWidth) {
-      // Position the cursor from the right of the block.
-      cursor.x = blockWidth - padding;
-    }
-
-    const positionBlock = function (block) {
-      const heightWidth = block.blockly_block.getHeightWidth();
-      const hasFrameSvg = !!block.blockly_block.functionalSvg_;
-      const frameSvgSize = hasFrameSvg ? 40 : 0;
-      const frameSvgTop = hasFrameSvg ? 35 : 0;
-      const frameSvgMargin = hasFrameSvg ? 16 : 0;
-
-      // If the block doesn't already have coordinates, use the cursor.
-      if (isNaN(block.x)) {
-        block.x = cursor.x;
-      } else if (blockSpace.RTL) {
-        // Position RTLs with coordinates from the left.
-        block.x = viewWidth - block.x;
-      }
-      // Adjust for Svg Frames for function definition blocks
-      if (!blockSpace.RTL) {
-        block.x += frameSvgMargin;
-      } else {
-        block.x -= frameSvgMargin;
-      }
-
-      if (isNaN(block.y)) {
-        block.y = cursor.y + frameSvgTop;
-        cursor.y +=
-          heightWidth.height + verticalSpaceBetweenBlocks + frameSvgSize;
-      }
-
-      block.blockly_block.moveTo(
-        new Blockly.utils.Coordinate(block.x, block.y)
-      );
-    };
-
-    blocks.sort(reorderBlocks);
-
-    blocks
-      .filter(function (block) {
-        return block.blockly_block.isVisible();
-      })
-      .forEach(positionBlock);
-
-    blocks
-      .filter(function (block) {
-        return !block.blockly_block.isVisible();
-      })
-      .forEach(positionBlock);
-
-    blockSpace.render();
     return blocks;
   };
 
   blocklyWrapper.Xml.blockSpaceToDom = blocklyWrapper.Xml.workspaceToDom;
-
-  // We don't want to save absolute position in the block XML
-  blocklyWrapper.Xml.blockToDomWithXY = blocklyWrapper.Xml.blockToDom;
-}
-
-// Compare function - Moves functional definitions to the end of a block list.
-function reorderBlocks(a, b) {
-  if (
-    a.blockly_block.type === 'procedures_defnoreturn' &&
-    b.blockly_block.type !== 'procedures_defnoreturn'
-  ) {
-    return 1; // Sort a after b.
-  } else if (
-    b.blockly_block.type === 'procedures_defnoreturn' &&
-    a.blockly_block.type !== 'procedures_defnoreturn'
-  ) {
-    return -1; // Sort a before b.
-  } else {
-    return 0; // Keep original order.
-  }
 }
